Program can connect to any number of slack servers
The configuration file now stores domains with their respective token in
a map[string]string.  If a command is invalid, the server will still
respond with "command not found" but if the command is valid, the server
will not send the response unless a valid domain is listed in the
config.

I will update the README.md to reflect this change in a later commit